add an encoding argument to utils.read no error for responses < 400
The encoding argument was also added to BasePeonyClient and will be
used each time a response is read. The argument was also added to
the PeonyClient.request() method.
By default encoding is set to None which means that it will use
aiohttp's default character encoding (right now they use chardet)
until now an error could be raised on responses 1xx and 3xx
I do not know any endpoint that would return these statuses but
I guess that doesn't mean an error should be raised