client: deprecate size_limit parameter
Instead of choosing the multipart upload only when the media exceeds a
given size the chunked parameter is just set to True by default.

This sounds more robust as I have no idea when Twitter decides to update
the limits and one user on IRC already reported needing to set the
chunked parameter to True by themselves. Also the API documentation says
"In the future, new features will only be supported for the chunked
upload endpoint."